Moved classes to include directory, updated 'require's
